+++ common/src/main/java/com/wynntils/core/persisted/config/ConfigManager.java
@@ -17,8 +17,8 @@ import com.wynntils.core.consumers.features.Translatable;
 import com.wynntils.core.consumers.overlays.DynamicOverlay;
 import com.wynntils.core.consumers.overlays.Overlay;
 import com.wynntils.core.consumers.overlays.OverlayManager;
+import com.wynntils.core.json.JsonManager;
 import com.wynntils.core.persisted.Persisted;
-import com.wynntils.core.persisted.json.JsonManager;
 import com.wynntils.core.persisted.upfixers.ConfigUpfixerManager;
 import com.wynntils.utils.JsonUtils;
 import com.wynntils.utils.mc.McUtils;
